Add confirmation functionnality, no email for now
There are two new columns in the table user, key and activated.
Activated is whether or not the account is activated, if it's set to
1, the account is activated and accessible, if not, it is not and the
user can't connect to his account.
There is a comment in loginController.php to show the link to confirm
the account of someone trying to connect with $error_msg destined to
testing purposes.